# sys_know/generated/openapi/schema.yaml
openapi: 3.0.3
info:
  title: TalentOS API
  version: 1.0.0
  description: Auto-generated API schema for testing and documentation
paths:
...
  /api/v1/passport/public/{token}/:
    get:
      operationId: v1_passport_public_retrieve
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: Success
        '404':
          description: Passport not found or link has expired.
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT